Check if config objects is being used by Service.Spec or Service.PreviousSpec before attempt to remove

+++ rootfs/promstack-housekeeping-agent.sh
@@ -1,15 +1,16 @@
-#!/bin/sh
+#!/bin/bash
 
+WHITELIST_CONFIGS=()
 LOGLEVEL=${LOGLEVEL:-info}
 HOUSEKEEPING_INTERVAL=${HOUSEKEEPING_INTERVAL:-300}
 DOCKER_STACK_NAMESPACE=${DOCKER_STACK_NAMESPACE:-promstack}
 
 logfmt() {
-	echo "ts=\"$(date +'%Y-%m-%dT%H:%M:%S%z')\" level=${LOGLEVEL:-info} $*"
+	echo "ts=\"$(date +'%Y-%m-%dT%H:%M:%S%z')\" level=${LOGFMTLEVEL:-info} $*"
 }
 logfmt_debug() {
 	if [ "${LOGLEVEL}" = "debug" ]; then
-		logfmt $*
+		LOGFMTLEVEL=debug logfmt $*
 	fi
 }
 
@@ -20,10 +21,17 @@ logfmt_oneline() {
 function docker_config_prune() {
 	local filter=$1
 	for cid in $(docker config ls -q --filter=label=${filter}); do
+		if [[ " ${WHITELIST_CONFIGS[@]} " =~ " ${cid} " ]]; then
+			logfmt_debug 'msg="Skip the housekeeping task on the Docker config object"' 'filter="label='${filter}'"' 'id="'$cid'"' 'status="in-used"'
+			continue
+		fi
+
+		local cname=$(docker config inspect $cid --format '{{.Spec.Name}}')
+
 		if docker config rm $cid > /dev/null 2>&1; then
-			logfmt 'msg="Perform housekeeping on Docker config object"' 'filter="label='${filter}'"' 'id="'$cid'"' 'status="removed"'
+			logfmt 'msg="Perform housekeeping on Docker config object"' 'filter="label='${filter}'"' 'id="'$cid'"' 'name="'$cname'"' 'status="removed"'
 		else
-			logfmt_debug 'msg="Perform housekeeping on Docker config object"' 'filter="label='${filter}'"' 'id="'$cid'"' 'status="skipped"'
+			logfmt_debug 'msg="Perform housekeeping on Docker config object"' 'filter="label='${filter}'"' 'id="'$cid'"' 'name="'$cname'"' 'status="skipped"'
 		fi
 		sleep 0.1
 	done
@@ -45,6 +53,27 @@ else
 
 		logfmt_debug 'msg="Starting the housekeeping task..."'
 		start=$(date +%s)
+
+		WHITELIST_CONFIGS=()
+		for sid in $(docker service ls -q); do
+			# ContainerSepc
+			TASKTEMPLATE_CONTAINERSPEC_CONFIGS_COUNT=$(docker service inspect $sid --format '{{.Spec.TaskTemplate.ContainerSpec.Configs | len}}')
+			if [ $TASKTEMPLATE_CONTAINERSPEC_CONFIGS_COUNT -gt 0 ]; then
+				TASKTEMPLATE_CONTAINERSPEC_CONFIGS=$(docker service inspect $sid --format '{{.Spec.TaskTemplate.ContainerSpec.Configs | json}}')
+				WHITELIST_CONFIGS+=($(echo ${TASKTEMPLATE_CONTAINERSPEC_CONFIGS} | jq -r '.[] | .ConfigID'))
+			fi
+
+			# PreviousSpec
+			# Check if service have PreviousSpec
+			if [ "$(docker service inspect $sid --format '{{.PreviousSpec}}')" = "<nil>" ]; then
+				continue
+			fi
+			PREVIOUS_TASKTEMPLATE_CONTAINERSPEC_CONFIGS_COUNT=$(docker service inspect $sid --format '{{.PreviousSpec.TaskTemplate.ContainerSpec.Configs | len}}')
+			if [ $PREVIOUS_TASKTEMPLATE_CONTAINERSPEC_CONFIGS_COUNT -gt 0 ]; then
+				TASKTEMPLATE_PREVIOUSSPEC_CONFIGS=$(docker service inspect $sid --format '{{.PreviousSpec.TaskTemplate.ContainerSpec.Configs | json}}')
+				WHITELIST_CONFIGS+=($(echo ${TASKTEMPLATE_PREVIOUSSPEC_CONFIGS} | jq -r '.[] | .ConfigID'))
+			fi
+		done
 
 		docker_config_prune "io.grafana.dashboard=true"
 		docker_config_prune "io.grafana.provisioning.alerting=true"
